+++ rebirth/R/extendr-wrappers.R
@@ -13,6 +13,10 @@ rebirth_handle_is_closed <- function(ptr) .Call(wrap__rebirth_handle_is_closed, 
 
 rebirth_available_backends <- function() .Call(wrap__rebirth_available_backends)
 
+rebirth_tokenize <- function(ptr, text, add_special, parse_special) .Call(wrap__rebirth_tokenize, ptr, text, add_special, parse_special)
+
+rebirth_detokenize <- function(ptr, ids) .Call(wrap__rebirth_detokenize, ptr, ids)
+
 rebirth_selftest_new_handle <- function() .Call(wrap__rebirth_selftest_new_handle)
 
 rebirth_selftest_panic <- function() .Call(wrap__rebirth_selftest_panic)